% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{college}
\alias{college}
\title{college dataset}
\format{
An object of class \code{data.frame} with 10000 rows and 3 columns.
}
\usage{
college
}
\description{
college dataset
}
\keyword{datasets}
