% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elections}
\alias{elections}
\title{elections dataset}
\format{
An object of class \code{data.frame} with 1092 rows and 3 columns.
}
\usage{
elections
}
\description{
elections dataset
}
\keyword{datasets}
