% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.1}
\alias{hw1_1.1}
\title{submit answer to homework 1 question 1.1}
\usage{
hw1_1.1(answer = c(NULL, "A", "B", "C"))
}
\arguments{
\item{answer}{a character string that is either 'A' 'B' or 'C' enter the letter corresponding to the correct statement in question 1.1}
}
\value{
summary of your current answer for question 1.1
}
\description{
submit answer to homework 1 question 1.1
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A wrong answer

# format answer to 1.1 like:
hw1_1.1(answer = 'B')

}

}
