% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.2.a}
\alias{hw1_1.2.a}
\title{submit answer to homework 1 question 1.2 part a}
\usage{
hw1_1.2.a(answer = NULL)
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 1.2 part a}
}
\value{
summary of your current answer for question 1.2 part a
}
\description{
submit answer to homework 1 question 1.2 part a
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different  wrong answer
# D: Another different wrong answer

# format answer to 1.2.a like:
hw1_1.2.a(answer = 'B')

}
}
