% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.2.b}
\alias{hw1_1.2.b}
\title{submit answer to homework 1 question 1.2 part b}
\usage{
hw1_1.2.b(answer = NULL)
}
\arguments{
\item{answer}{a character string that is letter'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 1.2 part b}
}
\value{
summary of your current answer for question 1.2 part b
}
\description{
submit answer to homework 1 question 1.2 part b
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different wrong answer
# D: Anther different wrong answer

# format answer to 1.2.b like:
hw1_1.2.b(answer = 'B')

}
}
