% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.2.c}
\alias{hw1_1.2.c}
\title{submit answer to homework 1 question 1.2 part c}
\usage{
hw1_1.2.c(answer = c(NULL, "A", "B", "C", "D"))
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 1.2 part c}
}
\value{
summary of your current answer for question 1.2 part c
}
\description{
submit answer to homework 1 question 1.2 part c
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different wrong answer
# D: Another different wrong answer

# format answer to 1.2.c like:
hw1_1.2.c(answer = 'B')

}
}
