% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.5.b}
\alias{hw1_1.5.b}
\title{submit answer to homework 1 question 1.5 part b}
\usage{
hw1_1.5.b(answer = c(NULL, TRUE, FALSE))
}
\arguments{
\item{answer}{a logical (TRUE/FALSE) indicating if the statement is TRUE or FALSE}
}
\value{
summary of your current answer for question 1.5 b
}
\description{
submit answer to homework 1 question 1.5 part b
}
\examples{
\donttest{
# These are made up and not related to actual answers
# Question 1.5 part b
#TURE or FALSE:
# A statement that is true
# format answer to 1.5 part b like:
hw1_1.5.b(TRUE)

# These are made up and not related to actual answers
# Question 1.5 part b
#TURE or FALSE:
# A statement that is false
# format answer to 1.5 part b like:
hw1_1.5.b(FALSE)

}
}
