% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_1.5.c}
\alias{hw1_1.5.c}
\title{submit answer to homework 1 question 1.5 part c}
\usage{
hw1_1.5.c(answer = NULL)
}
\arguments{
\item{answer}{a logical (TRUE/FALSE) indicating if the statement is TRUE or FALSE}
}
\value{
summary of your current answer for question 1.5 c
}
\description{
submit answer to homework 1 question 1.5 part c
}
\examples{
\donttest{
# These are made up and not related to actual answers
# Question 1.5 part c
#TURE or FALSE:
# A statement that is true
# format answer to 1.5 part c like:
hw1_1.5.c(TRUE)

# These are made up and not related to actual answers
# Question 1.5 part c
#TURE or FALSE:
# A statement that is false
# format answer to 1.5 part c like:
hw1_1.5.c(FALSE)

}
}
