% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_2.2}
\alias{hw1_2.2}
\title{submit answer to homework 1 question 2.2}
\usage{
hw1_2.2(answer = NULL)
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 2.2}
}
\value{
summary of your current answer for question 2.2
}
\description{
submit answer to homework 1 question 2.2
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different wrong answer
# D: Another different wrong answer

# format answer to 2.2 like:
hw1_2.2(answer = 'B')

}
}
