% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_2.4}
\alias{hw1_2.4}
\title{submit answer to homework 1 question 2.4}
\usage{
hw1_2.4(answer = c(NULL, "A", "B", "C"))
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 2.4}
}
\value{
summary of your current answer for question 2.4
}
\description{
submit answer to homework 1 question 2.4
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different wrong answer
# D: Another different wrong answer

# format answer to 2.4 like:
hw1_2.4(answer = 'B')

}
}
