% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{hw1_3.3}
\alias{hw1_3.3}
\title{submit answer to homework 1 question 3.3}
\usage{
hw1_3.3(answer = c(NULL, TRUE, FALSE))
}
\arguments{
\item{answer}{a logical (TRUE/FALSE) indicating if the statement made in question 3.3 is TRUE or FALSE}
}
\value{
summary of your current answer for question 3.3
}
\description{
submit answer to homework 1 question 3.3
}
\examples{
\donttest{
# These are made up and not related to actual answers for
# Question 3.3:
# A statement that is true.

# format answer to 3.3 like:
hw1_3.3(answer = TRUE)

}
}
