% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_3.2}
\alias{hw2_3.2}
\title{submit answer to homework 2 question 3.2}
\usage{
hw2_3.2(answer = c(NULL, "A", "B", "C", "D"))
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' or 'D'. For answer, enter the letter corresponding to the correct statement in question 3.2}
}
\description{
submit answer to homework 2 question 3.2
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different  wrong answer
# D: Another different wrong answer

# format answer to question  3.2 like:
hw2_3.2(answer = 'B')

}

}
