% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_3.3}
\alias{hw2_3.3}
\title{submit answer to homework 2 question 3.3}
\usage{
hw2_3.3(answer = NULL)
}
\arguments{
\item{answer}{a logical (TRUE/FALSE) indicating if the statement made in question 3.3 is TRUE or FALSE}
}
\description{
submit answer to homework 2 question 3.3
}
\examples{
\donttest{
# These are made up and not related to actual answers for
# Question 3.3:
# A statement that is true.

# format answer to 3.3 like:
hw2_3.3(answer = TRUE)

}
}
