% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_3.4}
\alias{hw2_3.4}
\title{submit answer to homework 2 question 3.4}
\usage{
hw2_3.4(answer = c(NULL, "A", "B", "C"))
}
\arguments{
\item{answer}{a character string that is either 'A' 'B' or 'C' enter the letter corresponding to the correct statement in question 3.4}
}
\value{
summary of your current answer for question 3.4
}
\description{
submit answer to homework 2 question 3.4
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A wrong answer

# format answer to 3.4 like:
hw2_3.4(answer = 'B')
}

}
