% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_4.1.a}
\alias{hw2_4.1.a}
\title{submit answer to homework 2 question 4.1 part a}
\usage{
hw2_4.1.a(est, se)
}
\arguments{
\item{est}{a numeric value of the point estimate of the SATE. Round to two decimal places.}

\item{se}{a numeric value of the standard error for the estimate of the SATE. Round to two decimal places.}
}
\description{
submit answer to homework 2 question 4.1 part a
}
\examples{
\donttest{

# Question 4.1 part a
# A statement that is true.

# format answer to question 4.1 part a like:

# a fake model
x <- rnorm(100, 0, 1)
z <- rbinom(100, 1, .5)
y1 <- x + 2 + rnorm(100)
y0 <- x + 2 + rnorm(100)
y <- ifelse(z == 1, y1, y0)
example <- data.frame(y, x, z)
fit <- lm(y ~ x + z, data = example)
# report rounded estimate and standard error
hw2_4.1.a(est = round(fit$coef[3], 2), se = round(sqrt(diag(vcov(fit)))[3], 2))

}

}
