% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_4.2.b}
\alias{hw2_4.2.b}
\title{submit answer to homework 2 question 4.2 part b}
\usage{
hw2_4.2.b(answer = c(NULL, TRUE, FALSE))
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 2 question 4.2 part b
}
\examples{
\donttest{

# Question 4.2 part b
# A statement that is true.

# format answer to question 4.2 part b like:
hw2_4.2.b(answer = TRUE)

# Question 4.2 part a
# A statement that is false.

# format answer to question 4.2 part b like:
hw2_4.2.b(answer = FALSE)

}

}
