% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{hw2_4.2.c}
\alias{hw2_4.2.c}
\title{submit answer to homework 2 question 4.2 part a}
\usage{
hw2_4.2.c(answer = NULL)
}
\arguments{
\item{answer}{a logical that is either TRUE or FALSE.}
}
\description{
submit answer to homework 2 question 4.2 part a
}
\examples{
\donttest{

# Question 4.2 part c
# A statement that is true.

# format answer to question 4.2 part c like:
hw2_4.2.c(answer = TRUE)

# Question 4.2 part c
# A statement that is false.

# format answer to question 4.2 part c like:
hw2_4.2.c(answer = FALSE)

}

}
