% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_1.3.a}
\alias{hw3_1.3.a}
\title{submit answer to homework 3 question 1.3 part a}
\usage{
hw3_1.3.a(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 3 question 1.3 part a
}
\examples{
\donttest{

# Question 1.3 part a
# TRUE or FALSE:
# A statement that is true.

# format answer to question 1.3 part a like:
hw3_1.3.a(answer = TRUE)

# Question 1.3 part a
# TRUE or FALSE:
# A statement that is false.

# format answer to question 1.3 part a like:
hw3_1.3.a(answer = FALSE)

}
}
