% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_2.4.b}
\alias{hw3_2.4.b}
\title{submit answer to homework 3 question 2.4 part b}
\usage{
hw3_2.4.b(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 3 question 2.4 part b
}
\examples{
\donttest{

# Question 2.4 part b
# TRUE or FALSE:
# A statement that is true.

# format answer to question 2.4 part b like:
hw3_2.4.b(answer = TRUE)

# Question 2.4 part b
# TRUE or FALSE:
# A statement that is false.

# format answer to question 2.4 part b like:
hw3_2.4.b(answer = FALSE)

}
}
