% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_2.5}
\alias{hw3_2.5}
\title{submit answer to homework 3 question 2.5}
\usage{
hw3_2.5(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 3 question 2.5
}
\examples{
\donttest{

# Question 2.5
# TRUE or FALSE:
# A statement that is true.

# format answer to question 2.5 like:
hw3_2.5(answer = TRUE)

# Question 2.5
# TRUE or FALSE:
# A statement that is false.

# format answer to question 2.5 like:
hw3_2.5(answer = FALSE)

}
}
