% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_3.2}
\alias{hw3_3.2}
\title{submit answer to homework 3 question 3.2}
\usage{
hw3_3.2(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 3 question 3.2
}
\examples{
\donttest{

# Question 3.2
# TRUE or FALSE:
# A statement that is true.

# format answer to question 3.2like:
hw3_3.2(answer = TRUE)

# Question 3.2
# TRUE or FALSE:
# A statement that is false.

# format answer to question 3.2 like:
hw3_3.2(answer = FALSE)

}
}
