% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_4.1}
\alias{hw3_4.1}
\title{submit answer to homework 3 question 4.1}
\usage{
hw3_4.1(est = NULL, se = NULL)
}
\arguments{
\item{est}{a numeric value of the point estimate of the SATE. Round to two decimal places.}

\item{se}{a numeric value of the standard error for the estimate of the SATE. Round to two decimal places.}
}
\description{
submit answer to homework 3 question 4.1
}
\examples{
\donttest{

# format answer to question 4.1 like:

# a fake model
x <- rnorm(100, 0, 1)
z <- rbinom(100, 1, .5)
y1 <- x + 2 + rnorm(100)
y0 <- x + 2 + rnorm(100)
y <- ifelse(z == 1, y1, y0)
example <- data.frame(y, x, z)
fit <- lm(y ~ x + z, data = example)
# report rounded estimate and standard error
hw3_4.1(est = round(fit$coef['z'], 2), se = round(sqrt(diag(vcov(fit)))['z'], 2))

}

}
