% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_1.1}
\alias{hw5_1.1}
\title{submit answer to homework 5 question 1.1}
\usage{
hw5_1.1(CACE = NULL)
}
\arguments{
\item{CACE}{a numeric value rounded to 2 decimal places. Report the value of the calculated CACE}
}
\description{
submit answer to homework 5 question 1.1
}
\examples{
\donttest{

# Question 1.1
# Lets say the true value of CACE is 2.343545235 (made up number)

# format answer to question 1.1 like:
hw5_1.1(CACE = 2.34)

}

}
