% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_1.2.b}
\alias{hw5_1.2.b}
\title{submit answer to homework 5 question 1.2 part a}
\usage{
hw5_1.2.b(SATE = NULL)
}
\arguments{
\item{SATE}{a numeric value rounded to 2 decimal places. Report the value of the calculated SATE for always takers}
}
\description{
submit answer to homework 5 question 1.2 part a
}
\examples{
\donttest{

# Question 1.2 part a
# Lets say the true value of SATE for always takers is 2.343545235 (made up number)

# format answer to question 1.3 like:
hw5_1.2.a(SATE = 2.34)

}

}
