% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_2.2.a}
\alias{hw5_2.2.a}
\title{submit answer to homework 5 question 2.2 part a}
\usage{
hw5_2.2.a(est = NULL)
}
\arguments{
\item{est}{a numeric value that is the estimated proportion of compliers. Round to two decimal places.}
}
\description{
submit answer to homework 5 question 2.2 part a
}
\examples{
\donttest{

# Question 2.2 part a
# Lets say the estimate proportion of compliers is .12345
# Remember proportions can not be greater than 1.
# you could submit your answer like this:
hw5_2.2.a(est = .12)

}

}
