% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_2.2.b}
\alias{hw5_2.2.b}
\title{submit answer to homework 5 question 2.2 part b}
\usage{
hw5_2.2.b(est = NULL)
}
\arguments{
\item{est}{a numeric value of the estimated as treated effect. Round to two decimal places.}
}
\description{
submit answer to homework 5 question 2.2 part b
}
\examples{
\donttest{

# Question 2.2 part b
# Lets say the estimate of your as treated analysis is 2.345345
# you could submit your answer like this:
hw5_2.2.b(est = 2.35)
}

}
