% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_2.2.d}
\alias{hw5_2.2.d}
\title{submit answer to homework 5 question 2.2 part d}
\usage{
hw5_2.2.d(est = NULL)
}
\arguments{
\item{est}{a numeric value of the estimated CACE from a Wald test. Round to two decimal places.}
}
\description{
submit answer to homework 5 question 2.2 part d
}
\examples{
\donttest{

# Question 2.2 part b
# Lets say the estimated CACE is 2.345345
# you could submit your answer like this:
hw5_2.2.d(est = 2.35)
}

}
