% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_3.1}
\alias{hw5_3.1}
\title{submit answer to homework 5 question 3.1}
\usage{
hw5_3.1(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 5 question 3.1
}
\examples{
\donttest{

# Question 3.1
# TRUE or FALSE:
# A statement that is true.

# format answer to question 3.1 like:
hw5_3.1(answer = TRUE)

# Question 3,1
# TRUE or FALSE:
# A statement that is false.

# format answer to question 3.1 like:
hw5_3.1(answer = FALSE)

}

}
