% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_3.2.a}
\alias{hw5_3.2.a}
\title{submit answer to homework 5 question 3.2 part a}
\usage{
hw5_3.2.a(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 5 question 3.2 part a
}
\examples{
\donttest{

# Question 3.2 part a
# TRUE or FALSE:
# A statement that is true.

# format answer to question 3.2 part a like:
hw5_3.2.a(answer = TRUE)

# Question 3.2 part a
# TRUE or FALSE:
# A statement that is false.

# format answer to question 3.2 part a like:
hw5_3.2.a(answer = FALSE)

}

}
