% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_3.2.d}
\alias{hw5_3.2.d}
\title{submit answer to homework 5 question 3.2 part d}
\usage{
hw5_3.2.d(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 5 question 3.2 part d
}
\examples{
\donttest{

# Question 3.2 part d
# TRUE or FALSE:
# A statement that is true.

# format answer to question 3.2 part d like:
hw5_3.2.d(answer = TRUE)

# Question 3.2 part d
# TRUE or FALSE:
# A statement that is false.

# format answer to question 3.2 part d like:
hw5_3.2.d(answer = FALSE)

}

}
