% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_3.2.e}
\alias{hw5_3.2.e}
\title{submit answer to homework 5 question 3.2 part e}
\usage{
hw5_3.2.e(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 5 question 3.2 part e
}
\examples{
\donttest{

# Question 3.2 part e
# TRUE or FALSE:
# A statement that is true.

# format answer to question 3.2 part e like:
hw5_3.2.e(answer = TRUE)

# Question 3.2 part e
# TRUE or FALSE:
# A statement that is false.

# format answer to question 3.2 part e like:
hw5_3.2.e(answer = FALSE)

}

}
