% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{hw5_4.1}
\alias{hw5_4.1}
\title{submit answer to homework 5 question 4.1}
\usage{
hw5_4.1(est = NULL, se = NULL)
}
\arguments{
\item{est}{a numeric value of the point estimate of the CACE. Round to two decimal places.}

\item{se}{a numeric value of the standard error for the estimate of the CACE. Round to two decimal places.}
}
\description{
submit answer to homework 5 question 4.1
}
\examples{
\donttest{

# Question 4.1
# Lets say the estimated CACE is 2.434453 (made up number) and the standard error is .8475492

# you could report your answer like this
hw5_4.1(est = 2.43, se = 8.5)
}

}
