% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_2.2.b}
\alias{hw6_2.2.b}
\title{submit answer to homework 6 question 2.2 part b}
\usage{
hw6_2.2.b(est, se)
}
\arguments{
\item{est}{the estimated SATE at the cut-point}

\item{se}{the standard error of the estimated SATE at the cut-point}
}
\description{
submit answer to homework 6 question 2.2 part b
}
