% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_3.1.a}
\alias{hw6_3.1.a}
\title{submit answer to homework 6 question 3.1 part a}
\usage{
hw6_3.1.a(est = NULL, se = NULL)
}
\arguments{
\item{est}{the estimated SATE at the cut-point}

\item{se}{the standard error of the estimated SATE at the cut-point}
}
\description{
submit answer to homework 6 question 3.1 part a
}
\examples{
\donttest{


# Question 3.1 part a
Lets assume the estimate at the cutoff is 4.99 with a standard error of .45869

# format answer to question 3.1 part a like:
hw6_3.1.a(est = 5.00, se = .46)

}

}
