% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_3.2.c}
\alias{hw6_3.2.c}
\title{submit answer to homework 6 question 3.2 part c}
\usage{
hw6_3.2.c(est, se)
}
\arguments{
\item{est}{the estimated SATE at the cut-point}

\item{se}{the standard error of the estimated SATE at the cut-point}
}
\description{
submit answer to homework 6 question 3.2 part c
}
