% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_3.4.c}
\alias{hw6_3.4.c}
\title{submit answer to homework 6 question 3.4 part c}
\usage{
hw6_3.4.c(est, se)
}
\arguments{
\item{est}{a numeric value representing the estimated SATE at the cut-point}

\item{se}{the standard error of the estimated SATE at the cut-point}
}
\description{
submit answer to homework 6 question 3.4 part c
}
\examples{
\donttest{


# Question 3.4 part c
Lets assume the estimate at the cutoff is 4.99 with a standard error of .45869

# format answer to question 3.4 part c like:
hw6_3.4.c(est = 5.00, se = .46)

}

}
