% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_3.5.b}
\alias{hw6_3.5.b}
\title{submit answer to homework 6 question 3.5 part b}
\usage{
hw6_3.5.b(est = NULL, lower.ci = NULL, upper.ci = NULL)
}
\arguments{
\item{est}{a numeric value representing the estimated SATE at the cut-point from an rdrobust analysis}

\item{lower.ci}{a numeric value representing the lower 95% CI from a rdrobust analysis}

\item{upper.ci}{a numeric value representing the upper 95% CI from a rdrobust analysis}
}
\description{
submit answer to homework 6 question 3.5 part b
}
\examples{
\donttest{


# Question 3.5 part b
# Lets assume the estimate at the cutoff is 4.99 with a lower CI 3.457 and an upper CI of 5.982

# format answer to question 3.5 part b like:
hw6_3.5.b(est = 5.00, lower.ci = 3.46, upper.ci = 5.98)

}

}
