% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_3.6}
\alias{hw6_3.6}
\title{submit answer to homework 6 question 3.6}
\usage{
hw6_3.6(answer = NULL)
}
\arguments{
\item{answer}{a character string that is letter 'A' 'B' 'C' 'D' 'E' or 'F'.}
}
\value{
summary of your current answer for question 3.6
}
\description{
submit answer to homework 6 question 3.6
}
\examples{
\donttest{
# These are made up and not related to actual answers
# A: A wrong answer
# B: The correct answer
# C: A different  wrong answer
# D: Another different wrong answer
# E: Yet another wrong answer
# F: A novel wrong answer

# format answer to 3.6 like:
hw6_3.6(answer = 'B')

}
}
