% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_5.2}
\alias{hw6_5.2}
\title{submit answer to homework 6 question 5.1}
\usage{
hw6_5.2(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 6 question 5.1
}
\examples{
\donttest{

# Question 5.2
# TRUE or FALSE:
# A statement that is true.

# format answer to question 5.2 like:
hw6_5.2(answer = TRUE)

# Question 5.2
# TRUE or FALSE:
# A statement that is false.

# format answer to question 5.2 like:
hw6_5.2(answer = FALSE)

}

}
