% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_7.1}
\alias{hw6_7.1}
\title{submit answer to homework 6 question 7.1}
\usage{
hw6_7.1(answer = NULL)
}
\arguments{
\item{answer}{a logical that is TRUE or FALSE}
}
\description{
submit answer to homework 6 question 7.1
}
\examples{
\donttest{

# Question 7.1
# TRUE or FALSE:
# A statement that is true.

# format answer to question 7.1 like:
hw6_7.1(answer = TRUE)

# Question 7.1
# TRUE or FALSE:
# A statement that is false.

# format answer to question 7.1 like:
hw6_7.1(answer = FALSE)

}

}
