% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_7.2}
\alias{hw6_7.2}
\title{submit answer to homework 6 question 7.2}
\usage{
hw6_7.2(est)
}
\arguments{
\item{est}{the estimated SATE at the cut-point#'}
}
\description{
submit answer to homework 6 question 7.2
}
