% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{hw6_7.3}
\alias{hw6_7.3}
\title{submit answer to homework 6 question 7.3}
\usage{
hw6_7.3(answer = NULL)
}
\arguments{
\item{answer}{a character vector that is 'Yes' or 'No'}
}
\description{
submit answer to homework 6 question 7.3
}
\examples{
\donttest{

# Question 7.3
# Yes or No:


# if answer is 'Yes' format answer to question 7.3 like:
hw6_7.3(answer = 'Yes')

# if answer is 'No' format answer to question 7.3 like:
hw6_7.3(answer = 'No')


}

}
