% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physics}
\alias{physics}
\title{physics dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 800 rows and 4 columns.
}
\usage{
physics
}
\description{
physics dataset
}
\keyword{datasets}
