% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_checks_hw2}
\alias{run_checks_hw2}
\title{Run checks on HW submissions}
\usage{
run_checks_hw2()
}
\description{
Run checks on HW submissions
}
\examples{
\donttest{
# After writing all your code and completing all answers run:
run_checks_hw2()

}
}
