% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_checks_hw5}
\alias{run_checks_hw5}
\title{Run homework checks for homework 5}
\usage{
run_checks_hw5()
}
\description{
Run homework checks for homework 5
}
\examples{
\donttest{
# After writing all your code and completing all answers run:
run_checks_hw5()

}

}
