% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_checks_hw6}
\alias{run_checks_hw6}
\title{Run homework checks for homework 6}
\usage{
run_checks_hw6()
}
\description{
Run homework checks for homework 6
}
\examples{
\donttest{
# After writing all your code and completing all answers run:
run_checks_hw6()

}

}
