% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sesame}
\alias{sesame}
\title{sesame dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 240 rows and 32 columns.
}
\usage{
sesame
}
\description{
sesame dataset
}
\keyword{datasets}
