% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{hw3_4.2}
\alias{hw3_4.2}
\title{submit answer to homework 3 question 4.2}
\usage{
hw3_4.2(answer = NULL)
}
\arguments{
\item{answer}{a chatacter that is either 'A' or 'B'}
}
\description{
submit answer to homework 3 question 4.2
}
\examples{
\donttest{


# Question 4.2
# A: the right answer.
# B: the wrong answer.

# format answer to question 4.2 like:
#hw3_4.2(answer = 'B')

}
}
